refactor(web): move componets to the root dir
