+++ web/components/Banner/index.ts
