+++ web/pages/_app.tsx
@@ -1,6 +1,6 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import Layout from "./components/Layout";
+import Layout from "../components/Layout";
 import Head from "next/head";
 
 function MyApp({ Component, pageProps }: AppProps) {
