+++ web/pages/charts.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Head from "next/head";
 import Link from "next/link";
-import { TextBanner } from "./components/Banner/Banner";
+import { TextBanner } from "../components/Banner/Banner";
 
 function Charts() {
   return (
