+++ web/pages/podfeed.tsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import React from "react";
-import { TextBanner } from "./components/Banner/Banner";
+import { TextBanner } from "../components/Banner/Banner";
 
 function Podfeed() {
   return (
